% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPP.R
\name{:=}
\alias{:=}
\title{:=}
\usage{
":="(lhs, rhs)
}
\arguments{
\item{lhs}{The value to be assigned to the left-hand argument (on the left)}

\item{rhs}{The value to be assigned to the right-hand argument (on the left)}
}
\description{
This function allows multiple assignments
}
\examples{
c(a, b):=c(3, 4)

}

