% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPP.R
\name{dirichlet.log.pdf}
\alias{dirichlet.log.pdf}
\title{dirichlet.log.pdf}
\usage{
dirichlet.log.pdf(K.probs, A)
}
\arguments{
\item{K.probs}{vector of probabilities}

\item{A}{vector of concentration parameters.}
}
\description{
The log of the probability density function for the Dirichlet distribution.
Returns the belief that the probabilities of K rival events are x_i
given that each event has been observed A_i - 1 times.
}

