% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{dirlnpdf}
\alias{dirlnpdf}
\title{dirlnpdf}
\usage{
dirlnpdf(X, A)
}
\arguments{
\item{X}{vector of probabilities}

\item{A}{vector of concentration parameters.}
}
\description{
The probability density function for the Dirichlet distribution.  Returns the belief that the probabilities of K rival events are x_i given that each event has been observed A_i -1 times.
}
\examples{
dirplndf(X,A)
}

