% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPP.R
\name{draw.L.given.N0}
\alias{draw.L.given.N0}
\title{draw.L.given.N0}
\usage{
draw.L.given.N0(N0, prior, EQUIV)
}
\arguments{
\item{N0}{Matrix containing the estimated baseline Poisson distribution for activity}

\item{prior}{Parameter values of a particular prior distribution}

\item{EQUIV}{Parameter sharing controls <- c(S1, S2):  S1 <- force sharing of delta (day effect) among days, S2 <- force sharing of eta (time of day) among days, Values: 1 (all days share), 2 (weekdays/weekends), 3 (none)}
}
\description{
Sample the L given N0
}

