% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPP.R
\name{draw.M.given.Z}
\alias{draw.M.given.Z}
\title{draw.M.given.Z}
\usage{
draw.M.given.Z(Z, prior)
}
\arguments{
\item{Z}{Binary vector indicationg the presence (1) or absence (0) of an event at every time slice}

\item{prior}{Parameter values of a particular prior distribution}
}
\description{
Sample the M given Z
}

