% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPP.R
\name{draw.Z.given.NLM}
\alias{draw.Z.given.NLM}
\title{draw.Z.given.NLM}
\usage{
draw.Z.given.NLM(N, L, M, priors)
}
\arguments{
\item{N}{Matrix of count data; axis 0 is the number of time intervals per day and axis 1 is the number of days in the data.}

\item{L}{Matrix containing the rate functions at every time slice}

\item{M}{Matrix containing the estimated transition probabilities for each iteration}

\item{priors}{List with parameter values of prior distributions}
}
\description{
Sample the given N, L, M
}

