% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPP.R
\name{generate.priors}
\alias{generate.priors}
\title{generate.priors}
\usage{
generate.priors(aL, bL, aD, aH, z00, z01, z10, z11, aE, bE, MODE)
}
\arguments{
\item{aL}{lambda0, baseline rate}

\item{bL}{lambda0, baseline rate}

\item{aD}{day effect dirichlet params}

\item{aH}{time of day effect dirichlet param}

\item{z00}{event process}

\item{z01}{event process}

\item{z10}{event process}

\item{aE}{gamma process}

\item{bE}{gamma process}

\item{MODE}{event process}
}
\description{
This function creates a list of prior parameters
}
\examples{
priors <- list()
priors$aL <- 1
priors$bL <- 1 #lambda0, baseline rate
priors$aD <- matrix(0, 1, 7)+5 #day effect dirichlet params
priors$aH <- matrix(0, nrow=48, ncol=7)+1 #time of day effect dirichlet param
priors$z01 <- .01*10000
priors$z00 <- .99*10000   #z(t) event process
priors$z01 <- .01*10000; priors$z00 <- .99*10000;     # z(t) event process
priors$z10 <- .25*10000; priors$z11 <- .75*10000;
priors$aE <- 5; priors$bE <- 1/3;  # gamma(t), or NBin, for event # process
priors$MODE <- 0;
}

