% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{logp}
\alias{logp}
\title{logp}
\usage{
logp(N, samples, priors, iter, EQUIV)
}
\arguments{
\item{N}{}

\item{samples}{}

\item{priors}{}

\item{iter}{}

\item{EQUIV}{}
}
\description{
Estimates the marginal likelihood of the data using the samples
}
\examples{
logp(N,samples,priors,iter,EQUIV)
}

