% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{nbinlnpdf}
\alias{nbinlnpdf}
\title{nbinlnpdf}
\usage{
nbinlnpdf(X, R, P)
}
\arguments{
\item{X}{}

\item{R}{}

\item{P}{}
}
\description{
The log probability density function for the negative binomial distribution.
}
\examples{
nbinlnpdf(X,R,P)
}

