% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPP.R
\name{neg.binomial.log.pdf}
\alias{neg.binomial.log.pdf}
\title{neg.binomial.log.pdf}
\usage{
neg.binomial.log.pdf(X, R, P)
}
\description{
The log probability density function for the negative binomial distribution.
}

