% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPP.R
\name{prob.L.given.N0}
\alias{prob.L.given.N0}
\title{prob.L.given.N0}
\usage{
prob.L.given.N0(L, N0, prior, EQUIV)
}
\arguments{
\item{L}{Matrix containing the rate functions at every time slice}

\item{N0}{Matrix containing the estimated baseline Poisson distribution for activity}

\item{prior}{Parameter values of a particular prior distribution}

\item{EQUIV}{Parameter sharing controls <- c(S1, S2):  S1 <- force sharing of delta (day effect) among days, S2 <- force sharing of eta (time of day) among days, Values: 1 (all days share), 2 (weekdays/weekends), 3 (none)}
}
\description{
This function evaluates p(L|N0)
}

