% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPP.R
\name{prob.M.given.Z}
\alias{prob.M.given.Z}
\title{prob.M.given.Z}
\usage{
prob.M.given.Z(M, Z, prior)
}
\arguments{
\item{M}{Matrix containing the estimated transition probabilities for each iteration}

\item{Z}{Binary vector indicationg the presence (1) or absence (0) of an event at every time slice}

\item{prior}{Parameter values of a particular prior distribution}
}
\description{
This function evaluates p(M|Z)
}

