% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPP.R
\name{prob.N.given.LM}
\alias{prob.N.given.LM}
\title{prob.N.given.LM}
\usage{
prob.N.given.LM(N, L, M, prior)
}
\arguments{
\item{N}{Matrix of count data; axis 0 is the number of time intervals per day and axis 1 is the number of days in the data.}

\item{L}{Matrix containing the rate functions at every time slice}

\item{M}{Matrix containing the estimated transition probabilities for each iteration}

\item{prior}{Parameter values of a particular prior distribution}
}
\description{
This function evaluates p(N|L, M)
}

