% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPP.R
\name{prob.N.given.LZ}
\alias{prob.N.given.LZ}
\title{prob.N.given.LZ}
\usage{
prob.N.given.LZ(N, L, Z, prior)
}
\arguments{
\item{N}{Matrix of count data; axis 0 is the number of time intervals per day and axis 1 is the number of days in the data.}

\item{L}{Matrix containing the rate functions at every time slice}

\item{Z}{Binary vector indicationg the presence (1) or absence (0) of an event at every time slice}

\item{prior}{Parameter values of a particular prior distribution}
}
\description{
This function evaluates p(N|L, Z)
}

