% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{repmat}
\alias{repmat}
\title{repmat}
\usage{
repmat(X, m, n)
}
\arguments{
\item{M}{a matrix}

\item{m}{new row dimension}

\item{n}{new column dimension}
}
\description{
This function replicates the matlab function repmat
}
\examples{
repmat(X,m,n)
}

