% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPP.R
\name{repmat}
\alias{repmat}
\title{repmat}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{Target matrix}

\item{m}{New row dimension}

\item{n}{New column dimension}
}
\description{
This function replicates the matlab function repmat
}

