% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{sensorMMPP}
\alias{sensorMMPP}
\title{sensorMMPP}
\usage{
sensorMMPP(N, priors, ITERS, EQUIV)
}
\arguments{
\item{N}{Matrix (Ntimes x 7*Nweeks) of count data (assumed starting Sunday) where Ntimes is the number of time intervals per day and Nweeks is the number of weeks in the data.}

\item{priors}{List with parameter values of prior distributions}

\item{ITERS}{Iteration controls: total # of iterations and # used for burn-in}

\item{EQUIV}{Parameter sharing controls = c(S1,S2):  S1 = force sharing of delta (day effect) among days, S2 = force sharing of eta (time of day) among days, Values: 1 (all days share), 2 (weekdays/weekends), 3 (none)}
}
\description{
This function provides the main MCMC inference engine
}
\examples{
priors<-list()
priors$aL=1
priors$bL=1 #lambda0, baseline rate
priors$aD=matrix(0,1,7)+5 #day effect dirichlet params
priors$aH=matrix(0,nrow=48,ncol=7)+1 #time of day effect dirichlet param
priors$z01=.01*10000
priors$z00=.99*10000   #z(t) event process
priors$z01 = .01*10000; priors$z00 = .99*10000;     # z(t) event process
priors$z10 = .25*10000; priors$z11 = .75*10000;
priors$aE = 5; priors$bE = 1/3;       # gamma(t), or NBin, for event # process
priors$MODE = 0;
sensorMMPP(N,priors,c(50,10),c(3,3))
}

