% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPP.R
\name{sensorMMPP}
\alias{sensorMMPP}
\title{sensorMMPP}
\usage{
sensorMMPP(N, priors = list(aL = 1, bL = 1, aD = matrix(0, 1, 7) + 5, aH =
  matrix(0, nrow = 48, ncol = 7) + 1, z00 = 0.99 * 10000, z01 = 0.01 * 10000,
  z10 = 0.25 * 10000, z11 = 0.75 * 10000, aE = 5, bE = 1/3, MODE = 0),
  ITERS = c(50, 10), EQUIV = c(3, 3))
}
\arguments{
\item{N}{Matrix of count data; axis 0 is the number of time intervals per day and axis 1 is the number of days in the data.}

\item{priors}{List with parameter values of prior distributions}

\item{ITERS}{Iteration controls: total # of iterations and # used for burn-in}

\item{EQUIV}{Parameter sharing controls <- c(S1, S2):
S1 <- force sharing of delta (day effect) among days,
S2 <- force sharing of eta (time of day) among days,
Values: 1 (all days share), 2 (weekdays/weekends), 3 (none)}
}
\description{
This function provides the main MCMC inference engine
}

